+++ solutions/ColumnFormatter/src/webparts/columnFormatter/components/Wizards/WizardCommon.ts
@@ -9,11 +9,12 @@ import { WizardDataBars } from './WizardDataBars';
 import { WizardDonut } from './WizardDonut';
 import { WizardFlow } from './WizardFlow';
 import { WizardMailTo } from './WizardMailTo';
+import { WizardMiniMap } from './WizardMiniMap';
 import { WizardNumberTending } from './WizardNumberTrending';
 import { WizardOverdue } from './WizardOverdue';
+import { WizardOverdueStatus } from './WizardOverdueStatus';
 import { WizardRoundImage } from './WizardRoundImage';
 import { WizardSeverity } from './WizardSeverity';
-import { WizardMiniMap } from './WizardMiniMap';
 import { WizardTwitter } from './WizardTwitter';
 
 //** Implement this interface to create your own wizard/template */
@@ -58,6 +59,7 @@ export const Wizards: Array<IWizard> = [
 	WizardMailTo,
 	WizardMiniMap,
 	WizardOverdue,
+	WizardOverdueStatus,
 	WizardRoundImage,
 	WizardSeverity,
 	WizardTwitter